% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pub.tabl.crea.R
\name{def.pub.tabl.crea}
\alias{def.pub.tabl.crea}
\title{Create publication table}
\usage{
def.pub.tabl.crea(data, pubWb, TablPub = NULL, log = NULL)
}
\arguments{
\item{data}{A list of Arrow datasets, each list element containing an arrow dataset as retrieved 
by arrow::open_dataset. These datasets contain time timeseries data from which the data for the 
publication table will be gathered from. Note that the datasets all must have the same number of rows.}

\item{pubWb}{A data from of the publication workbook, as generated by NEONprocIS.pub::def.read.pub.wb}

\item{TablPub}{Optional. Character value indicating the publication table to produce. If not input or 
the value is NULL, the entire pub workbook will be created. Note that duplicated field names in the 
pubWb (after filtering for the table of interest) will be removed.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A named list:
dataTabl: A data frame of the publication table. Note that the column order is determined by the 
          order of the rows in the publication workbook, not from the Rank field. This is intentional, 
          so as to avoid conflicts if there are duplicated field names or Rank values.
nameVarMtch: A character vector with the names of the variables in the pub table that were found in 
             the input datasets.
}
\description{
Definition function.
}
\examples{
Currently none
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[arrow]{dataset}

\link[NEONprocIS.pub]{def.read.pub.wb}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
