% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.read.pub.wb.R
\name{def.read.pub.wb}
\alias{def.read.pub.wb}
\title{Read Publication Workbook}
\usage{
def.read.pub.wb(NameFile, log = NULL)
}
\arguments{
\item{NameFile}{Character array. Name(s) (including relative or absolute path) of publication workbook file(s).}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame with all input publication workbooks combined.
}
\description{
Definition function. Read in one or more tab-delimited publication workbooks.
}
\examples{
pubWb <- NEONprocIS.base::def.read.pub.wb(NameFile=c('/scratch/mypubWb1.txt','/scratch/mypubWb2.txt'))
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
