% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.srf.term.qf.finl.R
\name{def.srf.term.qf.finl}
\alias{def.srf.term.qf.finl}
\title{Determine term of final quality flag given the term of the science review flag}
\usage{
def.srf.term.qf.finl(termSrf, log = NULL)
}
\arguments{
\item{termSrf}{Character vector. Term name(s) of the science review flag(s).}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A character vector with the term names of the corresponding final quality flags. Standalone
SRFs will return character(NA) for the resultant final quality flag.
}
\description{
Definition function. The science review flag has a specific naming convention based on the 
final quality flag to which it applies (see below). This function back-retrieves the term name of the
final quality flag given the tern name of the science review flag. Standalone science review
flags have no corresponding final quality flag. This scenario is also covered.
The naming convention for science review flags that modify a final quality flag is:
     <final quality flag term>SciRvw
The naming convention for standalone science review flags is sciRvwQf or <prefix>SciRvwQF.
}
\examples{
termSrf <- c("finalQFSciRvw", # modifies final quality flag
             "gWatSCondFinalQFSciRvw", # modifies final quality flag
             "sciRvwQF", # standalone SRF
             "tempSciRvwQF") # standalone SRF
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
